+++ dukes-payara/src/main/java/eu/agilejava/dukes/ApplicationConfig.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package eu.agilejava.snoop;
+package eu.agilejava.dukes;
 
 import java.util.HashSet;
 import java.util.Set;
